Use aliased shared_ptr instead of newGlobalBuffer
From http://en.cppreference.com/w/cpp/memory/shared_ptr/shared_ptr

template< class Y >
shared_ptr( const shared_ptr<Y>& r, element_type* ptr ) noexcept;

The aliasing constructor: constructs a shared_ptr which shares ownership
information with r, but holds an unrelated and unmanaged pointer ptr.
Even if this shared_ptr is the last of the group to go out of scope, it
will call the destructor for the object originally managed by r.
However, calling get() on this will always return a copy of ptr. It is
the responsibility of the programmer to make sure that this ptr remains
valid as long as this shared_ptr exists, such as in the typical use
cases where ptr is a member of the object managed by r or is an alias
(e.g., downcast) of r.get()

Emphasis on this part: "typical use cases where ptr is a member of the
object managed by r"